Use ptrdiff_t as iterator variable type in for loop

+++ src/solvers/SimInf_solver_mssm.c
@@ -82,7 +82,7 @@ SimInf_solver_mssm(
                  * and every node. Store the sum of the transition
                  * rates in each node in sum_t_rate. Moreover,
                  * initialize time in each node. */
-                for (int node = 0; node < m.Nn && !m.error; node++) {
+                for (ptrdiff_t node = 0; node < m.Nn && !m.error; node++) {
                     m.sum_t_rate[node] = 0.0;
                     for (int j = 0; j < m.Nt; j++) {
                         const double rate = (*m.tr_fun[j])(
@@ -117,7 +117,7 @@ SimInf_solver_mssm(
                 for (;!m.error;) {
                     /* (1) Handle internal epidemiological model,
                      * continuous-time Markov chain. */
-                    for (int node = 0; node < m.Nn && !m.error; node++) {
+                    for (ptrdiff_t node = 0; node < m.Nn && !m.error; node++) {
                         for (;;) {
                             double cum, rand, tau, delta = 0.0;
                             int tr;
@@ -222,7 +222,7 @@ SimInf_solver_mssm(
                      * each timestep e.g. update the infectious
                      * pressure variable. Moreover, update transition
                      * rates in nodes that are indicated for update */
-                    for (int node = 0; node < m.Nn && !m.error; node++) {
+                    for (ptrdiff_t node = 0; node < m.Nn && !m.error; node++) {
                         const int rc = m.pts_fun(
                             &m.v_new[node * m.Nd],
                             &m.u[node * m.Nc],
